Move CHC to own crate
