+++ crates/holochain/Cargo.toml
@@ -33,7 +33,6 @@ holo_hash = { version = "^0.4.0-dev.11", path = "../holo_hash", features = [
   "full",
 ] }
 holochain_cascade = { version = "^0.4.0-dev.20", path = "../holochain_cascade" }
-holochain_chc = { version = "^0.1.0-dev.1", path = "../holochain_chc" }
 holochain_conductor_api = { version = "^0.4.0-dev.20", path = "../holochain_conductor_api" }
 holochain_deepkey_dna = "0.0.8-dev.1"
 holochain_keystore = { version = "^0.4.0-dev.18", path = "../holochain_keystore", default-features = false }
@@ -86,7 +85,7 @@ thiserror = "1.0.22"
 tokio = { version = "1.36.0", features = ["full"] }
 tokio-stream = { version = "0.1", features = ["sync", "net"] }
 task-motel = "0.1.0"
-tracing = "0.1.37"
+tracing = "0.1"
 tracing-futures = "0.2.5"
 tracing-subscriber = "0.3.16"
 url = "2.4"
@@ -98,12 +97,13 @@ wasmer = "=4.3.1"
 wasmer-middlewares = "=4.3.1"
 opentelemetry_api = { version = "=0.20.0", features = ["metrics"] }
 
-# Dependencies for test_utils
+# Dependencies for test_utils / other optional deps
 arbitrary = { version = "1.0", features = ["derive"], optional = true }
 contrafact = { version = "0.2.0-rc.1", optional = true }
 diff = { version = "0.1", optional = true }
 hdk = { version = "^0.4.0-dev.14", path = "../hdk", optional = true }
 matches = { version = "0.1.8", optional = true }
+holochain_chc = { version = "^0.1.0-dev.1", path = "../holochain_chc", optional = true }
 holochain_test_wasm_common = { version = "^0.4.0-dev.14", path = "../test_utils/wasm_common", optional = true }
 kitsune_p2p_bootstrap = { version = "^0.3.0-dev.13", path = "../kitsune_p2p/bootstrap", optional = true }
 unwrap_to = { version = "0.1.0", optional = true }
@@ -189,7 +189,7 @@ path = "src/bin/holochain/main.rs"
 workspace = true
 
 [features]
-default = ["sqlite-encrypted", "tx2", "tx5", "metrics_influxive"]
+default = ["chc", "sqlite-encrypted", "tx2", "tx5", "metrics_influxive"]
 
 tx2 = ["kitsune_p2p/tx2"]
 tx5 = ["kitsune_p2p/tx5", "tx5-go-pion-turn"]
@@ -280,7 +280,7 @@ sqlite = [
 wasmer_debug_memory = ["holochain_wasmer_host/debug_memory"]
 
 # Enable chain head coordination
-chc = ["bytes", "reqwest", "holochain_conductor_api/chc", "holochain_types/chc"]
+chc = ["bytes", "reqwest", "holochain_conductor_api/chc", "holochain_chc"]
 
 # Transitional feature flag for code that is only ready when DPKI integration lands.
 dpki = []
