+++ crates/holochain/src/conductor/cell.rs
@@ -19,10 +19,10 @@ use tracing_futures::Instrument;
 use error::CellError;
 use holo_hash::*;
 use holochain_cascade::authority;
+use holochain_chc::ChcImpl;
 use holochain_conductor_api::ZomeCall;
 use holochain_nonce::fresh_nonce;
 use holochain_p2p::event::CountersigningSessionNegotiationMessage;
-use holochain_p2p::ChcImpl;
 use holochain_p2p::HolochainP2pDna;
 use holochain_sqlite::prelude::*;
 use holochain_state::host_fn_workspace::SourceChainWorkspace;
