+++ crates/holochain/src/conductor/chc.rs
@@ -1,7 +1,6 @@
 //! Types for Chain Head Coordination
 
 use holochain_keystore::MetaLairClient;
-use holochain_p2p::ChcImpl;
 use holochain_zome_types::prelude::*;
 use once_cell::sync::Lazy;
 use std::{collections::HashMap, sync::Arc};
