+++ crates/holochain/src/conductor/chc/chc_local.rs
@@ -1,5 +1,6 @@
 use std::sync::Arc;
 
+use holochain_chc::*;
 use holochain_keystore::MetaLairClient;
 use holochain_types::prelude::*;
