+++ crates/holochain_state/Cargo.toml
@@ -24,6 +24,7 @@ holo_hash = { version = "^0.4.0-dev.11", path = "../holo_hash", features = [
 ] }
 fallible-iterator = "0.3.0"
 hc_sleuth = { version = "^0.4.0-dev.20", path = "../hc_sleuth" }
+holochain_chc = { version = "^0.1.0-dev.1", path = "../holochain_chc" }
 holochain_keystore = { version = "^0.4.0-dev.18", path = "../holochain_keystore" }
 holochain_serialized_bytes = "=0.0.55"
 holochain_p2p = { version = "^0.4.0-dev.20", path = "../holochain_p2p" }
@@ -79,6 +80,7 @@ default = []
 fuzzing = ["holochain_types/fuzzing", "holochain_zome_types/fuzzing"]
 
 test_utils = [
+  "holochain_chc/test_utils",
   "holochain_keystore/test_utils",
   "holochain_types/test_utils",
   "holochain_zome_types/test_utils",
