+++ crates/holochain_state/src/source_chain.rs
@@ -15,8 +15,8 @@ use holo_hash::AgentPubKey;
 use holo_hash::DhtOpHash;
 use holo_hash::DnaHash;
 use holo_hash::HasHash;
+use holochain_chc::*;
 use holochain_keystore::MetaLairClient;
-use holochain_p2p::ChcImpl;
 use holochain_p2p::HolochainP2pDnaT;
 use holochain_sqlite::rusqlite::params;
 use holochain_sqlite::rusqlite::Transaction;
