+++ crates/holochain_types/src/lib.rs
@@ -18,7 +18,6 @@ pub mod activity;
 pub mod app;
 pub mod autonomic;
 pub mod chain;
-pub mod chc;
 pub mod combinators;
 pub mod db;
 pub mod db_cache;
